+++ exporter/exporterhelper/common.go
@@ -23,7 +23,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenthelper"
 	"go.opentelemetry.io/collector/config/configmodels"
-	"go.opentelemetry.io/collector/consumer/consumererror"
 )
 
 // ComponentSettings for timeout. The timeout applies to individual attempts to send data to the backend.
@@ -47,7 +46,7 @@ type request interface {
 	setContext(context.Context)
 	export(ctx context.Context) error
 	// Returns a new request that contains the items left to be sent.
-	onPartialError(consumererror.PartialError) request
+	onPartialError(error) request
 	// Returns the count of spans/metric points or log records.
 	count() int
 }