+++ exporter/exporterhelper/tracehelper.go
@@ -45,8 +45,8 @@ func newTracesRequest(ctx context.Context, td pdata.Traces, pusher PushTraces) r
 	}
 }
 
-func (req *tracesRequest) onPartialError(partialErr consumererror.PartialError) request {
-	return newTracesRequest(req.ctx, partialErr.GetTraces(), req.pusher)
+func (req *tracesRequest) onPartialError(err error) request {
+	return newTracesRequest(req.ctx, consumererror.GetTraces(err), req.pusher)
 }
 
 func (req *tracesRequest) export(ctx context.Context) error {