+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Remove ValidateConfig and add Validate on the Config struct (#2665)
 - Rename pdata Size to OtlpProtoSize (#2726)
 - Rename [Traces|Metrics|Logs]Consumer to [Traces|Metrics|Logs] (#2761)
+- Refactored `consumererror` package, eliminating `PartialError` type (#TBD)
 
 ## 💡 Enhancements 💡